+++ changelog.md
@@ -1,4 +1,9 @@
-## 5.1.2
+## 5.2.0
+* The Spawner Module has been rewritten!
+* Spawner Modifiers are now a JSON-based recipe system, with two inputs to any number of stat changes.
+* Removed the "Ignore Spawn Cap" modifier.
+* Added the "No AI" modifier.
+* Added the "Ignores Light" modifier.
 * Fixed a bug with the Enchantment Library where books above level 16 broke everything.
 * Made the enchantment library require Infused Hellshelves instead of standard.
 * Added the Library of Alexandria, which goes up to level 31.
@@ -6,6 +11,8 @@
 * Aikini: Updated chinese translation.
 * Re-Added compat for quark ancient tomes.
 * Fixed issues caused by disabling modules.
+* Fixed advancements being granted at the wrong time.
+* Added support for The One Probe.
 
 ## 5.1.1
 * Fixed a crash caused by a missing null check on LootingLevelEvent#getDamageSource